支持小数

+++ ouo/includes/ouoval.h
@@ -26,7 +26,7 @@ typedef struct ouoval {
     /* Basic
      * Error and Symbol types have some string data
      */
-    long num;
+    double num;
     char * err;
     char * sym;
     char * str;
@@ -72,7 +72,7 @@ ouoval * ouoval_err(const char* fmt, ...);
  *
  *  @return ouoval
  */
-ouoval * ouoval_num(long x);
+ouoval * ouoval_num(double x);
 
 /**
  *  @brief  Construct a pointer to a new Symbol ouoval
